- Use LauncherMain_t struct instead of writing wildly unreadable code - Fix Linux gmod_launcher build on (missing -ldl) - Fix Linux gmod_launcher not being output in INSTALL target - Fix Linux gmod_launcher code formatting - Rename CMake project gmod -> gmod_launcher - Rename Windows gmod_launcher.exe output to gmod.exe
