+++ CMakeLists.txt
@@ -72,7 +72,7 @@ add_subdirectory(html_chromium)
 add_subdirectory(chromium_process)
 add_subdirectory(example_host EXCLUDE_FROM_ALL)
 if("${CMAKE_SYSTEM_NAME}" STREQUAL "Linux")
-  add_subdirectory(gmod)
+  add_subdirectory(gmod_launcher)
 endif()
 
 PRINT_CEF_CONFIG()
